fix(frontend): record app update time at bundle startup
- Persist each version timestamp when the bundle first executes
- Keep build metadata in a pure shared module
- Verify later status-page visits preserve the original timestamp
